% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{animation_ease_in}
\alias{animation_ease_in}
\title{Eased in rotation}
\usage{
animation_ease_in(rdog, id = NULL, addTo = NULL, frames = Inf,
  framesPerCycle = 150, pause = 0, radiansPerCycle = tau,
  rotateAxis = "y", power = 2)
}
\arguments{
\item{rdog}{rdog object to add the animation to. Can be a character if called from a code block in shiny}

\item{id}{Id of the animation. You should set this if you want to be able
to check if this animation is running when getting inputs in a shiny app}

\item{addTo}{Which element should this animation be added to. If NULL, animation
will be set to the entire illustration.}

\item{frames}{How many frames should the animation run for. Usefult for animations triggered in shiny applications}

\item{framesPerCycle}{How many frames should a single eased in cycle take}

\item{radiansPerCycle}{How many radians should the object rotate per cycle. By default a single cycle will
cause a full rotation.}

\item{rotateAxis}{Axis of rotation}

\item{power}{Exponential power of the easing curve.}
}
\description{
Rotates the object smoothly
}
