% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{animation_rotate}
\alias{animation_rotate}
\title{Basic rotation}
\usage{
animation_rotate(rdog = NULL, id = NULL, addTo = NULL,
  frames = Inf, rotate = c(x = 0, y = 0, z = 0))
}
\arguments{
\item{rdog}{rdog object to add the animation to. Can be a character if called from a code block in shiny}

\item{id}{Id of the animation. You should set this if you want to be able
to check if this animation is running when getting inputs in a shiny app}

\item{addTo}{Which element should this animation be added to. If NULL, animation
will be set to the entire illustration.}

\item{frames}{How many frames should the animation run for. Usefult for animations triggered in shiny applications}

\item{rotate}{Rate of rotation per frame in radians.}
}
\description{
Rotates the object at a given speed.
}
