% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{binary_to_ascii_stl}
\alias{binary_to_ascii_stl}
\title{Convert binary STL to ASCII stl}
\usage{
binary_to_ascii_stl(file, output = NULL)
}
\arguments{
\item{file}{Input filepath}

\item{output}{output filepath}
}
\description{
Convert binary STL to ASCII stl
}
