% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{center_stl}
\alias{center_stl}
\title{Create the transform parameter to center an stl}
\usage{
center_stl(stl_bounds, exclude = NULL)
}
\arguments{
\item{stl_bounds}{output of \code{\link{get_stl_bounds}}}
}
\description{
Create the transform parameter to center an stl
}
