% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{copy_graph}
\alias{copy_graph}
\title{Copy an object and its children}
\usage{
copy_graph(rdog = NULL, id, what, ...)
}
\arguments{
\item{rdog}{rdog object where the element is in. Can be a character if called from a code block in shiny}

\item{id}{id of the element to be created}

\item{what}{id of the element to be copied}

\item{...}{Properties to be modified.}
}
\description{
Copy an object and its children
}
