% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{dog_metrics}
\alias{dog_metrics}
\title{Metrics}
\usage{
dog_metrics(api_key, application_key)
}
\arguments{
\item{api_key}{Your Datadog API key}

\item{application_key}{Your Application Key}
}
\description{
Return a list of all active metrics
}
\examples{
dog_metrics()
}

