% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{dog_query}
\alias{dog_query}
\title{Querying}
\usage{
dog_query(api_key, application_key, query, from_t, to_t, as_df = FALSE)
}
\arguments{
\item{api_key}{Your Datadog API key}

\item{application_key}{Your Application Key}

\item{query}{You search query}

\item{from_t}{Start of time range}

\item{to_t}{End of time range}

\item{as_df}{Boolean, returns xts when FALSE, data.fame when TRUE}
}
\description{
Querying
}
\examples{
\dontrun{dog_query('api_key', 'app_key', 1478872763, as.integer(Sys.time()), 'avg:system.cpu.idle{*}')}
}

