% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{dog_search}
\alias{dog_search}
\title{Search}
\usage{
dog_search(api_key, application_key, query)
}
\arguments{
\item{api_key}{Your Datadog API key}

\item{application_key}{Your Application Key}

\item{query}{Query string}
}
\description{
Search for entries from the last 24 hours
}
\examples{
\dontrun{dog_search('api_key', 'app_key', 'test')}
}

