% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_stl_bounds}
\alias{get_stl_bounds}
\title{Get boundaries of an stl file}
\usage{
get_stl_bounds(stl)
}
\arguments{
\item{stl}{A character with contents of an stl file}
}
\description{
Get boundaries of an stl file
}
