\name{rdog-package}
\alias{rdog-package}
\alias{rdog}
\docType{package}
\title{
rdog
}
\description{
A simple interface to Datadog.
}
\details{
\tabular{ll}{
Package: \tab rdog\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-04\cr
License: \tab GPL 3\cr
}
}
\author{
Alexis Le-Quoc

Maintainer: Alexis Le-Quoc <alq666@gmail.com>
}
\keyword{ package }