% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordGif.R
\name{record_gif}
\alias{record_gif}
\title{Record a gif from an illustration}
\usage{
record_gif(rdog, file = NULL, duration = 3)
}
\arguments{
\item{rdog}{rdog object}

\item{file}{output file path for the gif file}

\item{duration}{seconds to record}
}
\description{
Record a gif from an illustration
}
