% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordGif.R
\name{save_image}
\alias{save_image}
\title{Save an image from an illustration}
\usage{
save_image(rdog, file = NULL)
}
\arguments{
\item{rdog}{rdog object}

\item{file}{output file path for the image file}
}
\description{
Save an image from an illustration
}
