% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{scale_to_int}
\alias{scale_to_int}
\title{Scale a vector to lie between a set interval}
\usage{
scale_to_int(x, min, max)
}
\arguments{
\item{x}{vector to be scaled}

\item{min}{minimum value of the new interval}

\item{max}{maximum value of the new interval}
}
\description{
A utility function to scale vectors to specific intervals.
}
