% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{shape}
\alias{shape}
\alias{internal_shape}
\title{Common shape arguments}
\usage{
internal_shape(color = "#333", stroke = 1, fill = TRUE,
  closed = TRUE, visible = TRUE, backface = TRUE, front = c(z = 1))
}
\arguments{
\item{color}{Color of the shape. A color string.}

\item{stroke}{Width of the shape line. 0 will make it invisible}

\item{fill}{Logical. Fill the inner shape area}

\item{closed}{Logical. Should the first point and the last be connected}

\item{visible}{Logical. Is the shape visible.}

\item{backface}{Logical. Should backface be visible or a color string to set a different
color}

\item{front.}{Determine where the front of the shape is to decide rendering backface color.
A vector with named x, y, z elements.}
}
\description{
Common shape arguments
}
\keyword{internal}
