% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{shape_ellipse}
\alias{shape_ellipse}
\title{Ellipse}
\usage{
shape_ellipse(rdog = NULL, id = NULL, diameter = 1,
  width = diameter, height = diameter, quarters = 4,
  color = "#333", stroke = 1, fill = TRUE, closed = TRUE,
  visible = TRUE, backface = TRUE, front = c(z = 1), addTo = NULL,
  translate = c(x = 0, y = 0, z = 0), rotate = c(x = 0, y = 0, z = 0),
  scale = c(x = 1, y = 1, z = 1))
}
\arguments{
\item{rdog}{rdog object to add the shape to. Can be a character if called from a code block in shiny}

\item{id}{id of the shape. If NULL, a random id will be assigned}

\item{diameter}{Diameter of the circle}

\item{width, height}{Width and height of the ellipse. Overrides diameter}

\item{quarters}{How many quarters should be drawn. 4 draws a whole circle, 2
is a semi circle etc.}

\item{color}{Color of the shape. A color string.}

\item{stroke}{Width of the shape line. 0 will make it invisible}

\item{fill}{Logical. Fill the inner shape area}

\item{closed}{Logical. Should the first point and the last be connected}

\item{visible}{Logical. Is the shape visible.}

\item{backface}{Logical. Should backface be visible or a color string to set a different
color}

\item{front}{Determine where the front of the shape is to decide rendering backface color.
A vector with named x, y, z elements.}

\item{addTo}{Id of the parent object. If an rdog object is piped and
addTo is set to NULL, the default parent will be the illustration itself.}

\item{translate}{Position relative to the origin. Origin point is defined based
on \code{addTo} parameter. A vector with named x, y, z elements.}

\item{rotate}{Set rotation. Unit is radians. Use with \code{\link[base]{pi}} constant. A vector with named x, y, z elements.}

\item{scale}{Scale dimensons. Can be an unnamed integer or a vector with named x, y ,z elements}
}
\description{
Add an elipse to an illustration
}
