% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{shape_rect}
\alias{shape_rect}
\title{Rectangle}
\usage{
shape_rect(rdog = NULL, id = NULL, width = 1, height = 1,
  color = "#333", stroke = 1, fill = TRUE, closed = TRUE,
  visible = TRUE, backface = TRUE, front = c(z = 1), addTo = NULL,
  translate = c(x = 0, y = 0, z = 0), rotate = c(x = 0, y = 0, z = 0),
  scale = c(x = 1, y = 1, z = 1))
}
\arguments{
\item{rdog}{rdog object to add the shape to}

\item{id}{id of the shape. If NULL, a random id will be assigned}

\item{width, height}{Width and height of the rectangle.}

\item{color}{Color of the shape. A color string.}

\item{stroke}{Width of the shape line. 0 will make it invisible}

\item{fill}{Logical. Fill the inner shape area}

\item{closed}{Logical. Should the first point and the last be connected}

\item{visible}{Logical. Is the shape visible.}

\item{backface}{Logical. Should backface be visible or a color string to set a different
color}

\item{addTo}{Id of the parent object. If an rdog object is piped and
addTo is set to NULL, the default parent will be the illustration itself.}

\item{translate}{Position relative to the origin. Origin point is defined based
on \code{addTo} parameter. A vector with named x, y, z elements.}

\item{rotate}{Set rotation. Unit is radians. Use with \code{\link[base]{pi}} constant. A vector with named x, y, z elements.}

\item{scale}{Scale dimensons. Can be an unnamed integer or a vector with named x, y ,z elements}
}
\description{
Add a rectangle to an illustration
}
