% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlToShape.R
\name{stl_to_shape}
\alias{stl_to_shape}
\title{STL to shape}
\usage{
stl_to_shape(rdog = NULL, id = NULL, stl, colorAxis = "z",
  colorMin = "#FFFFFF", colorMax = "#000000", stroke = 1,
  addTo = NULL, translate = c(x = 0, y = 0, z = 0), rotate = c(x = 0,
  y = 0, z = 0), scale = c(x = 1, y = 1, z = 1))
}
\arguments{
\item{rdog}{rdog object to add the shape to. Can be a character if called from a code block in shiny}

\item{id}{id of the anchor. If NULL, a random id will be assigned}

\item{stl}{contents of an ascii stl file as a character}

\item{colorAxis}{axis to apply the color gradient}

\item{colorMin}{Starting color of the color gradient}

\item{colorMax}{Ending color of the color gradient}

\item{stroke}{Stroke to add to the polygons.}

\item{addTo}{Id of the parent object. If an rdog object is piped and
addTo is set to NULL, the default parent will be the illustration itself.}

\item{translate}{Position relative to the origin. Origin point is defined based
on \code{addTo} parameter. A vector with named x, y, z elements.}

\item{rotate}{Set rotation. Unit is radians. Use with \code{\link[base]{pi}} constant. A vector with named x, y, z elements.}

\item{scale}{Scale dimensons. Can be an unnamed integer or a vector with named x, y ,z elements}
}
\description{
Creates a shape from an ASCII stl
}
