% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update_property}
\alias{update_property}
\title{Update a property of an illustration element}
\usage{
update_property(rdog = NULL, elementId, ...)
}
\arguments{
\item{rdog}{rdog object where the element is in. Can be a character if called from a code block in shiny}

\item{elementId}{id of the element to be modified}

\item{...}{Properties to be modified.}
}
\description{
Update a property of an illustration element
}
