% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zfont.R
\name{zfont_font}
\alias{zfont_font}
\title{Create a zfont}
\usage{
zfont_font(rdog = NULL, id,
  font = system.file("fonts/Roboto-Regular.ttf", package = "rdog"))
}
\arguments{
\item{rdog}{rdog object to add the font to.}

\item{id}{id of the font}

\item{font}{file path to the ttf file that will be used}
}
\description{
Create a zfont that can be used to add text to an illustration
}
