% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{animation_none}
\alias{animation_none}
\title{No animation}
\usage{
animation_none(rdog = NULL, id = NULL)
}
\description{
Just establishes the animation loop without actually moving anything. This
is required if you called \code{\link{illustration}} with dragRotate = TRUE
}
